Refactored filepaths
