+++ src/test/java/com/nttdata/nttdatanotificationservice/service/WebHookServiceImplTest.java
@@ -5,8 +5,8 @@ import com.nttdata.nttdatanotificationservice.rocket.RocketChannelService;
 import com.nttdata.nttdatanotificationservice.sources.generic.models.GenericAlert;
 import com.nttdata.nttdatanotificationservice.sources.splunk.models.SplunkAlert;
 import com.nttdata.nttdatanotificationservice.sources.splunk.models.SplunkResult;
-import com.nttdata.nttdatanotificationservice.sources.WebHookParams;
-import com.nttdata.nttdatanotificationservice.sources.WebHookUrls;
+import com.nttdata.nttdatanotificationservice.configuration.WebHookParams;
+import com.nttdata.nttdatanotificationservice.configuration.WebHookUrls;
 import com.nttdata.nttdatanotificationservice.teams.TeamsChannelService;
 import org.junit.jupiter.api.*;
 import okhttp3.mockwebserver.MockResponse;
